// src/app/page/master/level-tingkat/level-tingkat.interface.ts
export interface LevelTingkat {

    kode?: any;
    kdLevelTingkat?: any;
    namaLevelTingkat?: string;
    reportDisplay?: string;
    kdLevelTingkatHead?: string;
    noUrut?: string;
    kdDepartemen?: any;
    kodeExternal?: string;
    namaExternal?: string;
    statusEnabled?: string;
    noRec?: string;
    version?: any;
    id?: any;
    label?: any;
    value?: any;
    namaLevelTingkatHead?: any;
}